screaper_backend/requirements.txt
python-dotenv==0.15.0
pyyaml==5.3.1
flashtext==2.7
w3lib==1.22.0
validators==0.18.2
testing.postgresql==1.3.0
gunicorn==20.0.4
Flask==1.1.2
flask_cors==3.0.10

pandas==1.1.4
sklearn
openpyxl
flask_sqlalchemy==2.4.4
SQLAlchemy-serializer==1.3.4

gevent==20.9.0
psycopg2-binary==2.8.6
pg8000==1.16.5
Pillow==8.1.1